+++ app/src/main/java/com/suka/superahorro/dbclasses/Item.kt
@@ -8,7 +8,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 class Item(var data: DbItem): Parcelable {
     constructor(name: String) : this(DbItem(name = name))
-    constructor(id: Int, name: String) : this(DbItem(id = id, name = name))
+    constructor(id: Long, name: String) : this(DbItem(id = id, name = name))
 
     fun toRef(): DbItemRef {
         return DbItemRef(id = data.id, name = data.name)
